requirements.txt
Mastodon.py==1.7.*
scipy==1.9.*